DAT-827-refactor-docapi-module
