--- app/src/app/docapi/docapi-add/docapi-add.component.html
@@ -1,2 +1,0 @@
-<cmdb-content-header title="Add new DocAPI document"></cmdb-content-header>
-<cmdb-docapi-settings-builder [mode]="mode"></cmdb-docapi-settings-builder>
