--- app/src/app/docapi/docapi-builder/docapi-builder-content-step/docapi-builder-content-step.component.html
@@ -1,14 +1,0 @@
-<div class="mt-3">
-  <div class="row">
-    <div class="col">
-      <form [formGroup]="contentForm" id="contentForm" class="needs-validation" novalidate autocomplete="off">
-        <div class="form-group">
-          <label for="docContentInput">Document Content: <span class="required">*</span></label>
-          <editor formControlName="template_data" id="docContentInput" aria-describedby="docContentInputHelp" [init]="editorConfig"></editor>
-          <small id="docLabelInputHelp" class="form-text text-muted">Design the content of your document</small>
-          <div class="clearfix"></div>
-        </div>
-      </form>
-    </div>
-  </div>
-</div>
