--- app/src/app/docapi/docapi-builder/docapi-builder-settings-step/docapi-builder-settings-step.component.html
@@ -1,56 +1,0 @@
-<div class="mt-3">
-  <div class="row">
-    <div class="col">
-      <form [formGroup]="settingsForm" id="settingsForm" class="needs-validation" novalidate autocomplete="off">
-        <div class="form-group">
-          <label>Active by default: <span class="required">*</span></label>
-          <div class="custom-control custom-switch">
-            <input type="checkbox" formControlName="active" class="custom-control-input" id="docActiveInput"/>
-            <label class="custom-control-label" for="docActiveInput"></label>
-          </div>
-        </div>
-
-        <div class="form-group">
-          <label for="docLabelInput">Label: <span class="required">*</span></label>
-          <input type="text" formControlName="label" class="form-control"
-                 [ngClass]="{ 'is-valid': label.valid && (label.dirty || label.touched),
-                 'is-invalid': label.invalid && (label.dirty || label.touched)}"
-                 id="docLabelInput" aria-describedby="docLabelInputHelp" required>
-          <div *ngIf="label.invalid && (label.dirty || label.touched)"
-               class="invalid-feedback">
-            <div class="float-right" *ngIf="label.errors.required">
-              Label is required.
-            </div>
-          </div>
-          <small id="docLabelInputHelp" class="form-text text-muted">Only a display name</small>
-          <div class="clearfix"></div>
-        </div>
-        <div class="form-group">
-          <label for="docNameInput">Name: <span class="required">*</span></label>
-          <input type="text" formControlName="name" class="form-control"
-                 [ngClass]="{ 'is-valid': name.valid && (name.dirty || name.touched),
-                 'is-invalid': name.invalid && (name.dirty || name.touched) }"
-                 id="docNameInput" readonly lowercase required>
-          <small id="docNameInputHelp" class="form-text text-muted">Document name must be lowercase unique</small>
-          <div *ngIf="name.invalid && (name.dirty || name.touched)"
-               class="invalid-feedback">
-            <div class="float-right" *ngIf="name.errors.required">
-              Name is required.
-            </div>
-            <div class="float-right" *ngIf="name.errors.nameAlreadyTaken">
-              A type with this name already exists.
-            </div>
-          </div>
-          <div class="clearfix"></div>
-        </div>
-        <div class="form-group">
-          <label for="docDescriptionInput">Description:</label>
-          <textarea class="form-control" formControlName="description" id="docDescriptionInput" rows="3"
-                    aria-describedby="docDescriptionInputHelp"></textarea>
-          <small id="docDescriptionInputHelp" class="form-text text-muted">Small description of the document</small>
-          <div class="clearfix"></div>
-        </div>
-      </form>
-    </div>
-  </div>
-</div>
