--- app/src/app/docapi/docapi-builder/docapi-builder-type-step-objectlist/docapi-builder-type-step-objectlist.component.html
@@ -1,15 +1,0 @@
-<form [formGroup]="typeParamForm" id="typeParamForm" class="needs-validation" novalidate autocomplete="off">
-  <div class="form-group">
-    <label for="docObjectTypeInput">Objectlist Type: <span class="required">*</span></label>
-    <ng-select id="docObjectTypeInput" [items]="objectTypeList"
-      bindLabel="label"
-      bindValue="public_id"
-      formControlName="type">
-      <ng-template ng-option-tmp let-item="item" let-index="index">
-        {{item.label}}
-      </ng-template>
-    </ng-select>
-    <small id="docTypeInputHelp" class="form-text text-muted">Choose your Object Type for your Template</small>
-    <div class="clearfix"></div>
-  </div>
-</form>
