--- app/src/app/docapi/docapi-edit/docapi-edit.component.html
@@ -1,2 +1,0 @@
-<cmdb-content-header title="Edit DocAPI document"></cmdb-content-header>
-<cmdb-docapi-settings-builder [docInstance]="docInstance" [mode]="mode"></cmdb-docapi-settings-builder>
