+++ app/src/app/modules/docapi/components/docapi-add/docapi-add.component.html
@@ -1,0 +1,2 @@
+<cmdb-content-header title="Add new DocAPI document"/>
+<cmdb-docapi-builder [mode]="mode"/>
