+++ app/src/app/modules/docapi/components/docapi-builder-style-step/docapi-builder-style-step.component.html
@@ -1,0 +1,22 @@
+<div class="mt-3">
+    <div class="row">
+        <div class="col">
+            <form [formGroup]="styleForm" id="styleForm" class="needs-validation" novalidate autocomplete="off">
+                <div class="form-group">
+                    <label for="docCustomCssInput">(Optional) Custom CSS:</label>
+                    <textarea
+                        class="form-control"
+                        formControlName="template_style"
+                        id="docCustomCssInput"
+                        rows="20"
+                        aria-describedby="docCustomCssInputHelp"
+                    ></textarea>
+                    <small id="docCustomCssInputHelp" class="form-text text-muted">
+                        Enter Custom CSS styling here
+                    </small>
+                    <div class="clearfix"></div>
+                </div>
+            </form>
+        </div>
+    </div>
+</div>
