+++ app/src/app/modules/docapi/components/docapi-builder-type-step-object/docapi-builder-type-step-object.component.html
@@ -1,0 +1,20 @@
+<form [formGroup]="typeParamForm" id="typeParamForm" class="needs-validation" novalidate autocomplete="off">
+    <div class="form-group">
+        <label for="docObjectTypeInput">Object Type: <span class="required">*</span></label>
+        <ng-select
+            id="docObjectTypeInput"
+            [items]="objectTypeList"
+            bindLabel="label"
+            bindValue="public_id"
+            formControlName="type"
+        >
+            <ng-template ng-option-tmp let-item="item" let-index="index">
+                {{item.label}}
+            </ng-template>
+        </ng-select>
+        <small id="docTypeInputHelp" class="form-text text-muted">
+            Choose your Object Type for your Template
+        </small>
+        <div class="clearfix"></div>
+    </div>
+</form>
