+++ app/src/app/modules/docapi/components/docapi-edit/docapi-edit.component.html
@@ -1,0 +1,2 @@
+<cmdb-content-header title="Edit DocAPI document"/>
+<cmdb-docapi-builder [docInstance]="docInstance" [mode]="mode"/>
